refactor(cfd-2d): Decompose momentum/boundary/mod.rs into leaf modules
Part of ATLAS-ARCH-007.
