+++ crates/cfd-2d/src/physics/momentum/boundary/mod.rs
@@ -87,81 +87,28 @@
 //! external forces and boundary fluxes, the total momentum $\int_\Omega \rho \mathbf{u} dV$
 //! is exactly conserved to machine precision.
 
+mod consistency;
 mod directional;
+mod higher_order;
+mod rotating;
+#[cfg(test)]
+mod tests;
+mod updater;
+
+pub use consistency::validate_boundary_consistency;
+pub use higher_order::apply_higher_order_wall_boundaries;
+pub use updater::MatrixUpdater;
 
 use super::solver::MomentumComponent;
-use crate::scalar;
 use crate::scalar::Cfd2dScalar;
-use cfd_core::error::{BoundaryErrorKind, Error, Result};
 use cfd_core::physics::boundary::BoundaryCondition;
-use cfd_math::sparse::SparseMatrixBuilder;
 use directional::{
     apply_east_boundary, apply_north_boundary, apply_south_boundary, apply_west_boundary,
 };
 use eunomia::FloatElement;
-use leto::geometry::Vector3;
 use leto::Array1;
 use std::collections::HashMap;
 use std::hash::BuildHasher;
-
-/// Abstract interface for sparse matrix topological assembly and zero-allocation value updates.
-///
-/// Permits generalized solvers to reuse matrix buffers regardless of whether they are initially
-/// constructing the CSR topology (via `SparseMatrixBuilder`) or performing in-place numerical
-/// updates on an existing structure (via `SparseMatrix`).
-pub trait MatrixUpdater<T> {
-    /// Inserts or aggregates `val` into the block matrix at location `(row, col)`.
-    ///
-    /// Implementations must uphold atomic accumulation commutativity for thread-parallel execution.
-    fn add_entry(&mut self, row: usize, col: usize, val: T) -> cfd_core::error::Result<()>;
-}
-
-impl<T: Cfd2dScalar + Copy> MatrixUpdater<T> for SparseMatrixBuilder<T> {
-    fn add_entry(&mut self, row: usize, col: usize, val: T) -> cfd_core::error::Result<()> {
-        self.add_entry(row, col, val)
-    }
-}
-
-impl<T: Cfd2dScalar + Copy> MatrixUpdater<T> for cfd_math::sparse::SparseMatrix<T> {
-    fn add_entry(&mut self, row: usize, col: usize, val: T) -> cfd_core::error::Result<()> {
-        let row_ptr = self.row_ptr().to_vec();
-        let col_indices = self.col_indices().to_vec();
-        let start = row_ptr[row];
-        let end = row_ptr[row + 1];
-        if let Ok(idx) = col_indices[start..end].binary_search(&col) {
-            self.values_mut()[start + idx] += val;
-        }
-        Ok(())
-    }
-}
-
-/// Apply rotating wall boundary condition: u_wall = ω × r
-/// where r is the position vector from center of rotation
-fn apply_rotating_wall_bc<T: Cfd2dScalar + Copy + FloatElement>(
-    component: MomentumComponent,
-    omega: &Vector3<T>,
-    center: &Vector3<T>,
-    grid: &crate::grid::StructuredGrid2D<T>,
-    idx: usize,
-) -> T {
-    let nx = grid.nx;
-    let i = idx % nx;
-    let j = idx / nx;
-
-    let cell_center = grid.cell_center(i, j).unwrap();
-    let x = cell_center[0];
-    let y = cell_center[1];
-
-    let r_x = x - center.x;
-    let r_y = y - center.y;
-
-    let omega_z = omega.z;
-
-    match component {
-        MomentumComponent::U => -omega_z * r_y,
-        MomentumComponent::V => omega_z * r_x,
-    }
-}
 
 /// Apply boundary conditions to momentum equation system
 pub fn apply_momentum_boundaries<T, S, M>(
@@ -179,7 +126,6 @@ where
     let nx = grid.nx;
     let ny = grid.ny;
     for (name, bc) in boundaries {
-        let name: &String = name;
         match name.as_str() {
             "west" => apply_west_boundary(matrix, rhs, bc, component, grid, nx, ny)?,
             "east" => apply_east_boundary(matrix, rhs, bc, component, grid, nx, ny)?,
@@ -190,414 +136,4 @@ where
     }
 
     Ok(())
-}
-
-/// Get the Dirichlet value for a boundary condition and component
-fn get_dirichlet_value<T: Cfd2dScalar + Copy + FloatElement>(
-    bc: &BoundaryCondition<T>,
-    component: MomentumComponent,
-) -> Option<T> {
-    match bc {
-        BoundaryCondition::Dirichlet {
-            value,
-            component_values,
-        } => {
-            if let Some(comps) = component_values {
-                let idx = match component {
-                    MomentumComponent::U => 0,
-                    MomentumComponent::V => 1,
-                };
-                if idx < comps.len() {
-                    comps[idx].or(Some(*value))
-                } else {
-                    Some(*value)
-                }
-            } else {
-                Some(*value)
-            }
-        }
-        BoundaryCondition::VelocityInlet { velocity } => {
-            let idx = match component {
-                MomentumComponent::U => 0,
-                MomentumComponent::V => 1,
-            };
-            Some(velocity[idx])
-        }
-        BoundaryCondition::Wall { wall_type } => match wall_type {
-            cfd_core::physics::boundary::WallType::NoSlip => Some(scalar::zero()),
-            cfd_core::physics::boundary::WallType::Moving { velocity } => {
-                let idx = match component {
-                    MomentumComponent::U => 0,
-                    MomentumComponent::V => 1,
-                };
-                Some(velocity[idx])
-            }
-            _ => None,
-        },
-        _ => None,
-    }
-}
-
-/// Validate boundary condition consistency
-pub fn validate_boundary_consistency<T, S>(
-    boundaries: &HashMap<String, BoundaryCondition<T>, S>,
-    _grid: &crate::grid::StructuredGrid2D<T>,
-) -> Result<()>
-where
-    T: Cfd2dScalar + Copy + FloatElement,
-    S: BuildHasher,
-{
-    for direction in &["north", "south", "east", "west"] {
-        if !boundaries.contains_key(*direction) {
-            return Err(Error::Boundary(BoundaryErrorKind::InvalidRegion(format!(
-                "Missing required boundary: {direction}"
-            ))));
-        }
-    }
-
-    for (name, bc) in boundaries {
-        if let BoundaryCondition::Periodic { partner } = bc {
-            if !boundaries.contains_key(partner) {
-                return Err(Error::Boundary(BoundaryErrorKind::InvalidRegion(format!(
-                    "Periodic partner '{partner}' not found for boundary '{name}'"
-                ))));
-            }
-            if let Some(partner_bc) = boundaries.get(partner) {
-                if let BoundaryCondition::Periodic { partner: p2 } = partner_bc {
-                    if p2 != name {
-                        return Err(Error::Boundary(BoundaryErrorKind::InvalidRegion(format!(
-                            "Periodic mismatch: '{name}' points to '{partner}', but '{partner}' points to '{p2}'"
-                        ))));
-                    }
-                } else {
-                    return Err(Error::Boundary(BoundaryErrorKind::InvalidRegion(format!(
-                        "Periodic partner '{partner}' is not periodic"
-                    ))));
-                }
-            }
-        }
-    }
-
-    let corners = [
-        ("west", "south"),
-        ("east", "south"),
-        ("west", "north"),
-        ("east", "north"),
-    ];
-
-    for (b1_name, b2_name) in corners {
-        let b1 = boundaries.get(b1_name).unwrap();
-        let b2 = boundaries.get(b2_name).unwrap();
-
-        for component in [MomentumComponent::U, MomentumComponent::V] {
-            if let (Some(v1), Some(v2)) = (
-                get_dirichlet_value(b1, component),
-                get_dirichlet_value(b2, component),
-            ) {
-                let diff = scalar::abs(v1 - v2);
-                let epsilon = T::default_epsilon() * scalar::from_f64(100.0);
-                if diff > epsilon {
-                    tracing::debug!(
-                        corner = %format!("{b1_name}-{b2_name}"),
-                        component = ?component,
-                        left = ?v1,
-                        right = ?v2,
-                        "Corner boundary values differ; retaining both side conditions"
-                    );
-                }
-            }
-        }
-    }
-
-    Ok(())
-}
-
-/// Apply higher-order boundary conditions for improved near-wall accuracy.
-/// Implements quadratic extrapolation for better velocity gradients near walls.
-pub fn apply_higher_order_wall_boundaries<T, S, M>(
-    matrix: &mut M,
-    rhs: &mut Array1<T>,
-    component: MomentumComponent,
-    boundaries: &HashMap<String, BoundaryCondition<T>, S>,
-    grid: &crate::grid::StructuredGrid2D<T>,
-) -> cfd_core::error::Result<()>
-where
-    T: Cfd2dScalar + Copy + FloatElement,
-    S: BuildHasher,
-    M: MatrixUpdater<T>,
-{
-    let nx = grid.nx;
-    let ny = grid.ny;
-
-    if let Some(BoundaryCondition::Wall {
-        wall_type: cfd_core::physics::boundary::WallType::NoSlip,
-    }) = boundaries.get("west")
-    {
-        if matches!(component, MomentumComponent::V) {
-            apply_higher_order_west_wall(matrix, rhs, component, grid, nx, ny)?;
-        }
-    }
-    if let Some(BoundaryCondition::Wall {
-        wall_type: cfd_core::physics::boundary::WallType::NoSlip,
-    }) = boundaries.get("east")
-    {
-        if matches!(component, MomentumComponent::V) {
-            apply_higher_order_east_wall(matrix, rhs, component, grid, nx, ny)?;
-        }
-    }
-    if let Some(BoundaryCondition::Wall {
-        wall_type: cfd_core::physics::boundary::WallType::NoSlip,
-    }) = boundaries.get("north")
-    {
-        if matches!(component, MomentumComponent::U) {
-            apply_higher_order_north_wall(matrix, rhs, component, grid, nx, ny)?;
-        }
-    }
-    if let Some(BoundaryCondition::Wall {
-        wall_type: cfd_core::physics::boundary::WallType::NoSlip,
-    }) = boundaries.get("south")
-    {
-        if matches!(component, MomentumComponent::U) {
-            apply_higher_order_south_wall(matrix, rhs, component, grid, nx, ny)?;
-        }
-    }
-
-    Ok(())
-}
-
-/// Quadratic extrapolation: u_0 = (4*u_1 - u_2)/3 for west wall
-fn apply_higher_order_west_wall<T: Cfd2dScalar + Copy + FloatElement, M: MatrixUpdater<T>>(
-    matrix: &mut M,
-    rhs: &mut Array1<T>,
-    _component: MomentumComponent,
-    _grid: &crate::grid::StructuredGrid2D<T>,
-    nx: usize,
-    ny: usize,
-) -> cfd_core::error::Result<()> {
-    let four: T = scalar::from_f64(4.0);
-    let three: T = scalar::from_f64(3.0);
-
-    // Horizontal walls own the corner nodes, so skip them here.
-    for j in 1..ny.saturating_sub(1) {
-        let idx_0 = j * nx;
-        let idx_1 = j * nx + 1;
-        let idx_2 = j * nx + 2;
-
-        matrix.add_entry(idx_0, idx_0, three)?;
-        matrix.add_entry(idx_0, idx_1, -four)?;
-        matrix.add_entry(idx_0, idx_2, scalar::one())?;
-        rhs[idx_0] = scalar::zero();
-    }
-
-    Ok(())
-}
-
-/// Quadratic extrapolation for east wall
-fn apply_higher_order_east_wall<T: Cfd2dScalar + Copy + FloatElement, M: MatrixUpdater<T>>(
-    matrix: &mut M,
-    rhs: &mut Array1<T>,
-    _component: MomentumComponent,
-    _grid: &crate::grid::StructuredGrid2D<T>,
-    nx: usize,
-    ny: usize,
-) -> cfd_core::error::Result<()> {
-    let four: T = scalar::from_f64(4.0);
-    let three: T = scalar::from_f64(3.0);
-
-    // Horizontal walls own the corner nodes, so skip them here.
-    for j in 1..ny.saturating_sub(1) {
-        let idx_0 = j * nx + nx - 1;
-        let idx_1 = j * nx + nx - 2;
-        let idx_2 = j * nx + nx - 3;
-
-        matrix.add_entry(idx_0, idx_0, three)?;
-        matrix.add_entry(idx_0, idx_1, -four)?;
-        matrix.add_entry(idx_0, idx_2, scalar::one())?;
-        rhs[idx_0] = scalar::zero();
-    }
-
-    Ok(())
-}
-
-/// Quadratic extrapolation for north wall
-fn apply_higher_order_north_wall<T: Cfd2dScalar + Copy + FloatElement, M: MatrixUpdater<T>>(
-    matrix: &mut M,
-    rhs: &mut Array1<T>,
-    _component: MomentumComponent,
-    _grid: &crate::grid::StructuredGrid2D<T>,
-    nx: usize,
-    ny: usize,
-) -> cfd_core::error::Result<()> {
-    let four: T = scalar::from_f64(4.0);
-    let three: T = scalar::from_f64(3.0);
-
-    for i in 0..nx {
-        let idx_0 = (ny - 1) * nx + i;
-        let idx_1 = (ny - 2) * nx + i;
-        let idx_2 = (ny - 3) * nx + i;
-
-        matrix.add_entry(idx_0, idx_0, three)?;
-        matrix.add_entry(idx_0, idx_1, -four)?;
-        matrix.add_entry(idx_0, idx_2, scalar::one())?;
-        rhs[idx_0] = scalar::zero();
-    }
-
-    Ok(())
-}
-
-/// Quadratic extrapolation for south wall
-fn apply_higher_order_south_wall<T: Cfd2dScalar + Copy + FloatElement, M: MatrixUpdater<T>>(
-    matrix: &mut M,
-    rhs: &mut Array1<T>,
-    _component: MomentumComponent,
-    _grid: &crate::grid::StructuredGrid2D<T>,
-    nx: usize,
-    _ny: usize,
-) -> cfd_core::error::Result<()> {
-    let four: T = scalar::from_f64(4.0);
-    let three: T = scalar::from_f64(3.0);
-
-    for i in 0..nx {
-        let idx_0 = i;
-        let idx_1 = nx + i;
-        let idx_2 = 2 * nx + i;
-
-        matrix.add_entry(idx_0, idx_0, three)?;
-        matrix.add_entry(idx_0, idx_1, -four)?;
-        matrix.add_entry(idx_0, idx_2, scalar::one())?;
-        rhs[idx_0] = scalar::zero();
-    }
-
-    Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::grid::StructuredGrid2D;
-    use cfd_core::physics::boundary::{BoundaryCondition, WallType};
-    use leto::geometry::Vector3;
-    use leto::Array1;
-    use std::collections::HashMap;
-
-    struct RecordingMatrix<T> {
-        entries: Vec<(usize, usize, T)>,
-    }
-
-    impl<T> RecordingMatrix<T> {
-        fn new() -> Self {
-            Self {
-                entries: Vec::new(),
-            }
-        }
-    }
-
-    impl<T: Cfd2dScalar + Copy> MatrixUpdater<T> for RecordingMatrix<T> {
-        fn add_entry(&mut self, row: usize, col: usize, val: T) -> cfd_core::error::Result<()> {
-            self.entries.push((row, col, val));
-            Ok(())
-        }
-    }
-
-    #[test]
-    fn higher_order_west_boundary_skips_shared_corners() {
-        let grid = StructuredGrid2D::new(4, 4, 0.0_f64, 1.0_f64, 0.0_f64, 1.0_f64)
-            .expect("grid creation failed");
-        let mut matrix = RecordingMatrix::new();
-        let mut rhs = Array1::from_elem([grid.nx * grid.ny], 0.0);
-        let boundaries = HashMap::from([(
-            "west".to_string(),
-            BoundaryCondition::Wall {
-                wall_type: WallType::NoSlip,
-            },
-        )]);
-
-        apply_higher_order_wall_boundaries(
-            &mut matrix,
-            &mut rhs,
-            MomentumComponent::V,
-            &boundaries,
-            &grid,
-        )
-        .expect("higher-order boundary application failed");
-
-        let top_left = (grid.ny - 1) * grid.nx;
-        let bottom_left = 0_usize;
-
-        assert!(
-            matrix.entries.iter().all(|(row, _, _)| *row != top_left),
-            "west higher-order boundary must not write the top-left corner"
-        );
-        assert!(
-            matrix.entries.iter().all(|(row, _, _)| *row != bottom_left),
-            "west higher-order boundary must not write the bottom-left corner"
-        );
-        assert!(
-            matrix.entries.iter().any(|(row, _, _)| *row == grid.nx),
-            "west higher-order boundary must still write interior west-wall rows"
-        );
-    }
-
-    #[test]
-    fn north_wall_owns_the_top_corners() {
-        let grid = StructuredGrid2D::new(4, 4, 0.0_f64, 1.0_f64, 0.0_f64, 1.0_f64)
-            .expect("grid creation failed");
-        let mut matrix = RecordingMatrix::new();
-        let mut rhs = Array1::from_elem([grid.nx * grid.ny], 0.0);
-        let boundaries = HashMap::from([
-            (
-                "north".to_string(),
-                BoundaryCondition::Wall {
-                    wall_type: WallType::Moving {
-                        velocity: Vector3::new(1.0, 0.0, 0.0),
-                    },
-                },
-            ),
-            (
-                "south".to_string(),
-                BoundaryCondition::Wall {
-                    wall_type: WallType::NoSlip,
-                },
-            ),
-            (
-                "west".to_string(),
-                BoundaryCondition::Wall {
-                    wall_type: WallType::NoSlip,
-                },
-            ),
-            (
-                "east".to_string(),
-                BoundaryCondition::Wall {
-                    wall_type: WallType::NoSlip,
-                },
-            ),
-        ]);
-
-        apply_higher_order_wall_boundaries(
-            &mut matrix,
-            &mut rhs,
-            MomentumComponent::U,
-            &boundaries,
-            &grid,
-        )
-        .expect("higher-order boundary application failed");
-        apply_momentum_boundaries(
-            &mut matrix,
-            &mut rhs,
-            MomentumComponent::U,
-            &boundaries,
-            &grid,
-        )
-        .expect("boundary application failed");
-
-        let top_left = (grid.ny - 1) * grid.nx;
-        let top_right = top_left + grid.nx - 1;
-        let bottom_left = 0_usize;
-        let bottom_right = grid.nx - 1;
-
-        assert_eq!(rhs[top_left], 1.0);
-        assert_eq!(rhs[top_right], 1.0);
-        assert_eq!(rhs[bottom_left], 0.0);
-        assert_eq!(rhs[bottom_right], 0.0);
-    }
 }
